# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Arborist
  version: 2.4.0
  description: authorization microservice to handle ABAC based on configured policies
  license:
    name: 'Apache 2.0'
    url: 'https://github.com/calypr/arborist'
...
paths:
...
  /auth/request:
    post:
      tags:
        - auth
      description: >-
        Ask arborist to check authorization on a user request. A client of the
        Gen3 auth stack sends the user's JWT to this endpoint, along with the
        information about the user's specific request. The terminology used to
        describe user requests is flexible and arborist does not "nail it down";
        however our usage so far is to use the permission model to capture the
        basic set of operations one might expect in ABAC (read/write, etc.)
        combined with the client service, and the path for the resource the user
        is trying to access.


        If using a list of `requests`, the response is positive if the user
        has access to ALL items in the list.

        If the given JWT has `azp` field, the permission of
        the corresponding client will be also checked; only when both the user
        and the client have permission can the response be positive. 


        Note: Under `body` we can also check authorization of
        a user given their username, using the field `user_id` instead of
        using `token`. This means that anyone can check anyone's authorization
        if they have their username. For this reason, this API is not meant to
        be exposed publicly. Arborist does not currently have
        authorization for the management of the authorization in the system.
        The service itself does not provide a way to expose only certain parts
        of the API to super admins vs admins.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestBody'
      responses:
        200:
          description: >-
            Arborist successfully returned an authorization decision. NOTE
            that a 200 status DOES NOT indicate authorization, only that the
            input was valid. The `"auth"` field in the response indicates
            authorization (this is different from
            the `/auth/proxy` endpoint behavior).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRequestResponse'
        400:
          description: >-
            The input was somehow invalid; for example, a given resource does
            not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
  /auth/proxy:
    get:
      tags:
        - auth
      description: >-
        Ask arborist to check authorization on a user request. A client of the
        Gen3 auth stack sends the user's JWT to this endpoint, along with the
        information about the user's specific request. The terminology used to
        describe user requests is flexible and arborist does not "nail it down";
        however our usage so far is to use the permission model to capture the
        basic set of operations one might expect in ABAC (read/write, etc.)
        combined with the client service, and the path for the resource the user
        is trying to access.


        If the given JWT has `azp` field, the permission of
        the corresponding client will be also checked; only when both the user
        and the client have permission can the response be positive.
      parameters:
        - in: query
          name: resource
          required: true
          schema:
            type: string
        - in: query
          name: service
          required: true
          schema:
            type: string
        - in: query
          name: method
          required: true
          schema:
            type: string
      responses:
        200:
          description: >-
            Arborist successfully returned an authorization decision. NOTE
            that a 200 status DOES indicate authorization (this is different from
            the `/auth/request` endpoint behavior).
        400:
          description: >-
            The input was somehow invalid; for example, a given resource does
            not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
        401:
          description: >-
            The user is not logged in.
        403:
          description: >-
            The user does not have access.
  /auth/resources:
    get:
      tags:
        - auth
      description: >-
        Given a user token in the Authorization header, return a list of
        resources which are accessible to the user (using any action) according
        to the policies listed in the token. This includes the resources available 
        to members of the `anonymous` and `logged-in` groups, which apply to all users 
        (anonymous) and to all logged-in users (logged-in).


        If the specified user is not recognized by arborist, this endpoint returns
        ONLY the resources available to members of the `anonymous` and `logged-in` groups.


        If no username is provided (i.e., if no token is provided in the Authorization header)
        this endpoint returns ONLY the resources available to members of the `anonymous` group.


        If the `tags` query string parameter is passed, this endpoint returns the resource
        tags for the resources, not the resources themselves.
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
        - in: query
          name: tags
          required: false
          schema:
            type: string
          description: Instructs the endpoint to return resource tags instead of resources.
      responses:
        200:
          description: >-
            Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResourcesResponse'
        401:
          description: >-
            Authorization header or token failed to validate (authentication error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthenticated'
    post:
      tags:
        - auth
      description: >-
        Given a user token in the body of the request, return a list of
        resources which are accessible to the user (using any action) according
        to the policies listed in the token. This includes the resources available 
        to members of the `anonymous` and `logged-in` groups, which apply to all users 
        (anonymous) and to all logged-in users (logged-in).


        If the specified user is not recognized by arborist, this endpoint returns
        ONLY the resources available to members of the `anonymous` and `logged-in` groups.


        If the `tags` query string parameter is passed, this endpoint returns the resource
        tags for the resources, not the resources themselves.
      parameters:
        - in: query
          name: tags
          required: false
          schema:
            type: string
          description: Instructs the endpoint to return resource tags instead of resources.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthResourcesRequestBody'
      responses:
        200:
          description: >-
            Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResourcesResponse'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
        401:
          description: >-
            Token failed to validate (authentication error)
...
  /resource/{resourcePath}:
    parameters:
      - in: path
        name: resourcePath
        required: true
        schema:
          type: string
        allowReserved: true
        description: >-
          The full path for a resource, which includes slashes. For example, if
          a resource was created which has the path `/a/b/c`, then the endpoint
          `/resource/a/b/c` can now be used to access this resource.
    get:
      tags:
        - resource
      description: Read the resource given by the path
      responses:
        200:
          description: JSON representation of the specified resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        404:
          description: no resource exists with the given `resourcePath`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    post:
      tags:
        - resource
      description: >-
        Add a new subresource, *underneath* an existing resource. The path *is
        allowed to be empty*; when `resourcePath` is empty, the resource is
        created under the root resource, as `/new_resource`.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceInput'
      parameters:
        - in: query
          name: p
          description: >-
            Ordinarily arborist will return an error if parent resources of a
            resource you're trying to create do not exist yet. If the `p`
            parameter is included, it behaves like `mkdir -p` and creates all
            parent resources as necessary.
          required: false
      responses:
        201:
          description: JSON representation of successfully-created resource
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    $ref: '#/components/schemas/Resource'
        404:
          description: no resource exists with the given `resourcePath`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
    put:
      tags:
        - resource
      description: >-
        Overwrite an existing resource. This endpoint requires a fully-formed
        resource model (and cannot patch over individual fields on the existing
        resources). If the specified resource doesn't exist, it will be created.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceInput'
      parameters:
        - in: query
          name: p
          description: >-
            Ordinarily arborist will return an error if parent resources of a
            resource you're trying to create do not exist yet. If the `p`
            parameter is included, it behaves like `mkdir -p` and creates all
            parent resources as necessary.
          required: false
        - in: query
          name: merge
          description: >-
            Ordinarily arborist will overwrite the resource by deleting
            existing subresources. If the `merge` parameter is included,
            those existing subresources are not deleted.
          required: false
      responses:
        201:
          description: JSON representation of successfully-updated resource
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    $ref: '#/components/schemas/Resource'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
    delete:
      tags:
        - resource
      responses:
        204:
          description: resource successfully deleted
        404:
          description: no resource exists with the given `resourcePath`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /role:
    get:
      tags:
        - role
      description: List all the roles registered in arborist.
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
    post:
      tags:
        - role
      description: Create a new role
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        201:
          description: Success; returns JSON representation of created role
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    $ref: '#/components/schemas/Role'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
  /role/{roleID}:
    parameters:
      - in: path
        name: roleID
        required: true
        schema:
          type: string
        description: The ID for a role registered in arborist.
    get:
      tags:
        - role
      description: >-
        Output the information for a specific role.
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        404:
          description: no role exists with the given `roleID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    put:
      tags:
        - role
      description: >-
        Overwrite an existing role with new content. This endpoint requires a
        fully-formed role (and cannot patch over individual fields on the
        existing resources).
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        201:
          description: Success; returns JSON representation of updated role
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    $ref: '#/components/schemas/Role'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
        404:
          description: no role exists with the given `roleID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    patch:
      tags:
        - role
      description: >-
        Append information to an existing role. The contents of all the fields
        provided in a `PATCH` request are appended to the existing fields on
        this role.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        201:
          description: Success; returns JSON representation of updated role
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    $ref: '#/components/schemas/Role'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
        404:
          description: no role exists with the given `roleID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    delete:
      tags:
        - role
      description: Delete this role.
      responses:
        204:
          description: role successfully deleted
        404:
          description: no role exists with the given `roleID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /policy:
    get:
      tags:
        - policy
      description: >-
        List just the IDs for all policies which have been created in arborist.
      parameters:
        - in: path
          name: expand
          required: false
          schema:
            type: boolean
          description: Whether to return detailed roles instead of only role IDs (disabled by default). If enabled, 'roles' will replace 'role_ids' in the returned data.
      responses:
        200:
          description: list of resources
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Policy'
    post:
      tags:
        - policy
      description: Create a new policy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        201:
          description: Success; returns JSON representation of created policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    $ref: '#/components/schemas/Policy'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
  /policy/{policyID}:
    parameters:
      - in: path
        name: policyID
        required: true
        schema:
          type: string
        description: The ID for a policy registered in arborist.
    get:
      tags:
        - policy
      description: >-
        Output the information for a specific policy.
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        404:
          description: no role exists with the given `policyID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    put:
      tags:
        - policy
      description: >-
        Overwrite an existing policy with new content. This endpoint requires a
        fully-formed policy (and cannot patch over individual fields on the
        existing resources).
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        201:
          description: Success; returns JSON representation of updated policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    $ref: '#/components/schemas/Policy'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
        404:
          description: no policy exists with the given `policyID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    patch:
      tags:
        - policy
      description: >-
        Append information to an existing policy. The contents of all the fields
        provided in a `PATCH` request are appended to the existing fields on
        this policy.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policy'
      responses:
        201:
          description: Success; returns JSON representation of updated policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    $ref: '#/components/schemas/Policy'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
        404:
          description: no policy exists with the given `policyID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
    delete:
      tags:
        - policy
      description: Delete this policy.
      responses:
        204:
          description: policy successfully deleted
        404:
          description: no policy exists with the given `policyID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /bulk/policy:
    put:
      tags:
        - policy
      description: >-
        Overwrite existing policies with new content. This endpoint requires a list of
        fully-formed policies (and cannot patch over individual fields on the 
        existing resources)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Policies'
      responses:
        201:
          description: Success; returns JSON representation of updated policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    $ref: '#/components/schemas/Policies'
        400:
          description: invalid input (missing fields or fields have incorrect types)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserError'
        404:
          description: no policy exists with the given `policyID`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
...
  /user/{username}/bulk/policy:
    parameters:
      - in: path
        name: username
        required: true
        schema:
          type: string
        description: the username for a user registered in arborist
    post:
      tags:
        - user
      description: >-
        Grant list of policies to a user. The policy must exist already.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantUserPolicies'
      responses:
        204:
          description: successful granted policies
        404:
          description: user not found, or policy not found
...
components:
  schemas:
...
    AuthRequestResponse:
      type: object
      properties:
        auth:
          type: boolean
    AuthResourcesRequestBody:
      type: object
      properties:
        user:
          type: object
          properties:
            token:
              type: string
              example: 'eyJhbGciOiJFUzI1NiIsImtpZCI6IjE2In0[...]'
          required:
            - token
      required:
        - user
    AuthResourcesResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            type: string
          example: ['/programs/DEV/projects/test', '/programs/foo/projects/bar']
    UserError:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              description: an explanation of the error that occurred
            code:
              type: integer
              description: the HTTP error code
      example:
        error:
          message: "input resource is missing the following required fields: ..."
          code: 400
...
    Policies:
      type: array
      description: list of policies
      items:
        $ref: '#/components/schemas/Policy'
...
    GrantUserPolicies:
      type: array
      description: list of policies for a user
      items:
        $ref: '#/components/schemas/GrantUserPolicy'
...
    Unauthenticated:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              example: failed to validate JWT
            code:
              type: integer
              example: 401
...